Fix Logger

+++ app/Specials/Logger.php
@@ -105,7 +105,7 @@ class Logger
         $this->contents .= '[' . date('d.m.y H:i:s') . '] ' . print_r($message, true) . "\n";
 
         // Write to immediate log
-        if ($this->tmpPath) {
+        if ($this->tmpPath && file_exists($this->tmpPath)) {
             file_put_contents($this->tmpPath, $this->contents);
         }
     }
@@ -125,10 +125,6 @@ class Logger
 
         // Write to each path
         foreach ($this->paths as $path) {
-            if (!file_exists($path)) {
-                continue;
-            }
-
             if (!$logFile = fopen($path, 'a')) {
                 continue;
             }
